+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 All notable changes to `blokctl` will be documented in this file.
 
-## 0.2.2 - WIP
+## 0.3.0 - 2026-03-17
+- Adding `folder:create`: Create a folder (with optional parent folder by slug or ID)
 - Adding `workflow:stage-show`: Show details of a workflow stage by name or ID
 - Adding `workflows:list`: List workflows and their stages (lookup stage IDs by name)
 - Adding `story:workflow-change`: Change the workflow stage of a story